Remove commented code & split code
